requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
redis==5.0.4
kafka-python==2.0.2
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
requests==2.32.3
rapidfuzz==3.9.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
locust==2.29.1
